sentinel-images/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
PyYAML==6.0.2
torch==2.3.1
torchvision==0.18.1
pillow==10.4.0
numpy==2.1.2
matplotlib==3.9.2
schedule==1.2.2
google-cloud-storage==2.18.2
